refactor: nest ConfirmInfoRow as discussed with @jpuri
